Added Concurrent Server Slot functionality and Began Celery Integration
Added Server slot functionality, which will help to limit the amount of servers that can be started simultaneously on the site. To prevent users from overloading the hoster.

In order to introduce automated background processes for purposes such as cleanup and health checks, the decision to integrate Celery with the Django project was made.

Currently this is being aimed to be introduced via WSL on a Windows Debian installation, Celery itself will be communicating with a Debian instance of Redis.